DPL: move analysis topology helper methods to a separate class
On their way to a separate plugin
