+++ Framework/Core/include/Framework/CommonDataProcessors.h
@@ -19,29 +19,8 @@
 
 namespace o2::framework
 {
-
-class DataOutputDirector;
-
-struct OutputTaskInfo {
-  uint32_t id;
-  std::string name;
-};
-
-struct OutputObjectInfo {
-  uint32_t id;
-  std::vector<std::string> bindings;
-};
-} // namespace o2::framework
-extern template class std::vector<o2::framework::OutputObjectInfo>;
-extern template class std::vector<o2::framework::OutputTaskInfo>;
-namespace o2::framework
-{
 /// Helpers to create a few general data processors
 struct CommonDataProcessors {
-  /// Match all inputs of kind ATSK and write them to a ROOT file,
-  /// one root file per originating task.
-  static DataProcessorSpec getOutputObjHistSink(std::vector<OutputObjectInfo> const& objmap,
-                                                std::vector<OutputTaskInfo> const& tskmap);
   /// Given the list of @a danglingInputs @return a DataProcessor which does
   /// a binary dump for all the dangling inputs matching the Timeframe
   /// lifetime. @a unmatched will be filled with all the InputSpecs which are
@@ -54,10 +33,6 @@ struct CommonDataProcessors {
   /// @fixme: for now only the dangling inputs are forwarded.
   static DataProcessorSpec getGlobalFairMQSink(std::vector<InputSpec> const& danglingInputs);
 
-  /// writes inputs of kind AOD to file
-  static DataProcessorSpec getGlobalAODSink(std::shared_ptr<DataOutputDirector> dod,
-                                            std::vector<InputSpec> const& outputInputs);
-
   /// @return a dummy DataProcessorSpec which requires all the passed @a InputSpec
   /// and simply discards them. @a rateLimitingChannelConfig is the configuration
   /// for the rate limiting channel, if any required.
@@ -67,4 +42,4 @@ struct CommonDataProcessors {
 
 } // namespace o2::framework
 
-#endif // o2_framework_CommonDataProcessors_H_INCLUDED
+#endif // O2_FRAMEWORK_COMMONDATAPROCESSORS_H_
